+++ packages/google_maps_flutter/google_maps_flutter_android/android/src/test/java/io/flutter/plugins/googlemaps/ConvertTest.java
@@ -4,6 +4,15 @@
 
 package io.flutter.plugins.googlemaps;
 
+import static io.flutter.plugins.googlemaps.Convert.HEATMAP_DATA_KEY;
+import static io.flutter.plugins.googlemaps.Convert.HEATMAP_GRADIENT_COLORS_KEY;
+import static io.flutter.plugins.googlemaps.Convert.HEATMAP_GRADIENT_COLOR_MAP_SIZE_KEY;
+import static io.flutter.plugins.googlemaps.Convert.HEATMAP_GRADIENT_KEY;
+import static io.flutter.plugins.googlemaps.Convert.HEATMAP_GRADIENT_START_POINTS_KEY;
+import static io.flutter.plugins.googlemaps.Convert.HEATMAP_ID_KEY;
+import static io.flutter.plugins.googlemaps.Convert.HEATMAP_MAX_INTENSITY_KEY;
+import static io.flutter.plugins.googlemaps.Convert.HEATMAP_OPACITY_KEY;
+import static io.flutter.plugins.googlemaps.Convert.HEATMAP_RADIUS_KEY;
 import static org.junit.Assert.fail;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.never;
@@ -17,9 +26,14 @@ import android.graphics.Color;
 import android.graphics.Paint;
 import android.os.Build;
 import android.util.Base64;
+import androidx.annotation.NonNull;
 import com.google.android.gms.maps.model.BitmapDescriptor;
 import com.google.android.gms.maps.model.LatLng;
 import com.google.maps.android.clustering.algo.StaticCluster;
+import com.google.maps.android.geometry.Point;
+import com.google.maps.android.heatmaps.Gradient;
+import com.google.maps.android.heatmaps.WeightedLatLng;
+import com.google.maps.android.projection.SphericalMercatorProjection;
 import io.flutter.plugins.googlemaps.Convert.BitmapDescriptorFactoryWrapper;
 import io.flutter.plugins.googlemaps.Convert.FlutterInjectorWrapper;
 import java.io.ByteArrayInputStream;
@@ -324,6 +338,128 @@ public class ConvertTest {
     fail("Expected an IllegalArgumentException to be thrown");
   }
 
+  private static final SphericalMercatorProjection sProjection = new SphericalMercatorProjection(1);
+
+  @Test()
+  public void ConvertToWeightedLatLngReturnsCorrectData() {
+    final double intensity = 3.3;
+    final Object data = List.of(List.of(1.1, 2.2), intensity);
+    final Point point = sProjection.toPoint(new LatLng(1.1, 2.2));
+
+    final WeightedLatLng result = Convert.toWeightedLatLng(data);
+
+    Assert.assertEquals(point.x, result.getPoint().x, 0);
+    Assert.assertEquals(point.y, result.getPoint().y, 0);
+    Assert.assertEquals(intensity, result.getIntensity(), 0);
+  }
+
+  @Test()
+  public void ConvertToWeightedDataReturnsCorrectData() {
+    final double intensity = 3.3;
+    final List<Object> data = List.of(List.of(List.of(1.1, 2.2), intensity));
+    final Point point = sProjection.toPoint(new LatLng(1.1, 2.2));
+
+    final List<WeightedLatLng> result = Convert.toWeightedData(data);
+
+    Assert.assertEquals(1, result.size());
+    Assert.assertEquals(point.x, result.get(0).getPoint().x, 0);
+    Assert.assertEquals(point.y, result.get(0).getPoint().y, 0);
+    Assert.assertEquals(intensity, result.get(0).getIntensity(), 0);
+  }
+
+  @Test()
+  public void ConvertToGradientReturnsCorrectData() {
+    final int color1 = 0;
+    final int color2 = 1;
+    final int color3 = 2;
+    final List<Object> colorData = List.of(color1, color2, color3);
+    final double startPoint1 = 0.0;
+    final double startPoint2 = 1.0;
+    final double startPoint3 = 2.0;
+    List<Object> startPointData = List.of(startPoint1, startPoint2, startPoint3);
+    final int colorMapSize = 3;
+    final Map<String, Object> data =
+        Map.of(
+            HEATMAP_GRADIENT_COLORS_KEY, colorData,
+            HEATMAP_GRADIENT_START_POINTS_KEY, startPointData,
+            HEATMAP_GRADIENT_COLOR_MAP_SIZE_KEY, colorMapSize);
+
+    final Gradient result = Convert.toGradient(data);
+
+    Assert.assertEquals(3, result.mColors.length);
+    Assert.assertEquals(color1, result.mColors[0]);
+    Assert.assertEquals(color2, result.mColors[1]);
+    Assert.assertEquals(color3, result.mColors[2]);
+    Assert.assertEquals(3, result.mStartPoints.length);
+    Assert.assertEquals(startPoint1, result.mStartPoints[0], 0);
+    Assert.assertEquals(startPoint2, result.mStartPoints[1], 0);
+    Assert.assertEquals(startPoint3, result.mStartPoints[2], 0);
+    Assert.assertEquals(colorMapSize, result.mColorMapSize);
+  }
+
+  @Test()
+  public void ConvertInterpretHeatmapOptionsReturnsCorrectData() {
+    final double intensity = 3.3;
+    final List<Object> dataData = List.of(List.of(List.of(1.1, 2.2), intensity));
+    final Point point = sProjection.toPoint(new LatLng(1.1, 2.2));
+
+    final int color1 = 0;
+    final int color2 = 1;
+    final int color3 = 2;
+    final List<Object> colorData = List.of(color1, color2, color3);
+    final double startPoint1 = 0.0;
+    final double startPoint2 = 1.0;
+    final double startPoint3 = 2.0;
+    List<Object> startPointData = List.of(startPoint1, startPoint2, startPoint3);
+    final int colorMapSize = 3;
+    final Map<String, ?> gradientData =
+        Map.of(
+            HEATMAP_GRADIENT_COLORS_KEY, colorData,
+            HEATMAP_GRADIENT_START_POINTS_KEY, startPointData,
+            HEATMAP_GRADIENT_COLOR_MAP_SIZE_KEY, colorMapSize);
+
+    final double maxIntensity = 4.4;
+    final double opacity = 5.5;
+    final int radius = 6;
+    final String idData = "heatmap_1";
+
+    final Map<String, Object> data =
+        Map.of(
+            HEATMAP_DATA_KEY,
+            dataData,
+            HEATMAP_GRADIENT_KEY,
+            gradientData,
+            HEATMAP_MAX_INTENSITY_KEY,
+            maxIntensity,
+            HEATMAP_OPACITY_KEY,
+            opacity,
+            HEATMAP_RADIUS_KEY,
+            radius,
+            HEATMAP_ID_KEY,
+            idData);
+
+    final MockHeatmapBuilder builder = new MockHeatmapBuilder();
+    final String id = Convert.interpretHeatmapOptions(data, builder);
+
+    Assert.assertEquals(1, builder.getWeightedData().size());
+    Assert.assertEquals(point.x, builder.getWeightedData().get(0).getPoint().x, 0);
+    Assert.assertEquals(point.y, builder.getWeightedData().get(0).getPoint().y, 0);
+    Assert.assertEquals(intensity, builder.getWeightedData().get(0).getIntensity(), 0);
+    Assert.assertEquals(3, builder.getGradient().mColors.length);
+    Assert.assertEquals(color1, builder.getGradient().mColors[0]);
+    Assert.assertEquals(color2, builder.getGradient().mColors[1]);
+    Assert.assertEquals(color3, builder.getGradient().mColors[2]);
+    Assert.assertEquals(3, builder.getGradient().mStartPoints.length);
+    Assert.assertEquals(startPoint1, builder.getGradient().mStartPoints[0], 0);
+    Assert.assertEquals(startPoint2, builder.getGradient().mStartPoints[1], 0);
+    Assert.assertEquals(startPoint3, builder.getGradient().mStartPoints[2], 0);
+    Assert.assertEquals(colorMapSize, builder.getGradient().mColorMapSize);
+    Assert.assertEquals(maxIntensity, builder.getMaxIntensity(), 0);
+    Assert.assertEquals(opacity, builder.getOpacity(), 0);
+    Assert.assertEquals(radius, builder.getRadius());
+    Assert.assertEquals(idData, id);
+  }
+
   private InputStream buildImageInputStream() {
     Bitmap fakeBitmap = Bitmap.createBitmap(1, 1, Bitmap.Config.ARGB_8888);
     ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
@@ -354,5 +490,58 @@ public class ConvertTest {
     String base64Image = Base64.encodeToString(pngBytes, Base64.DEFAULT);
 
     return base64Image;
+  }
+}
+
+class MockHeatmapBuilder implements HeatmapOptionsSink {
+  private List<WeightedLatLng> weightedData;
+  private Gradient gradient;
+  private double maxIntensity;
+  private double opacity;
+  private int radius;
+
+  public List<WeightedLatLng> getWeightedData() {
+    return weightedData;
+  }
+
+  public Gradient getGradient() {
+    return gradient;
+  }
+
+  public double getMaxIntensity() {
+    return maxIntensity;
+  }
+
+  public double getOpacity() {
+    return opacity;
+  }
+
+  public int getRadius() {
+    return radius;
+  }
+
+  @Override
+  public void setWeightedData(@NonNull List<WeightedLatLng> weightedData) {
+    this.weightedData = weightedData;
+  }
+
+  @Override
+  public void setGradient(@NonNull Gradient gradient) {
+    this.gradient = gradient;
+  }
+
+  @Override
+  public void setMaxIntensity(double maxIntensity) {
+    this.maxIntensity = maxIntensity;
+  }
+
+  @Override
+  public void setOpacity(double opacity) {
+    this.opacity = opacity;
+  }
+
+  @Override
+  public void setRadius(int radius) {
+    this.radius = radius;
   }
 }