+++ packages/shared_preferences/shared_preferences/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 2.4.0
 
 * Adds shared preferences devtools extension
+* Updates minimum supported SDK version to Flutter 3.22/Dart 3.4.
+* Updates minimum supported SDK version to Flutter 3.19/Dart 3.3.
 
 ## 2.3.1